Update README.md
I apologize for so many edits, still trying to figure out the path I want for the program, these are the current things I need. I still have more things I need I ultimately want this to be an aesthetically pleasing program you can edit till your heart is content, maybe an image to the background, I want to create it as easily as possible to change the theme without prior knowledge of coding. But we will see.